+++ lib/html/search_result_types/Artwork.rb
@@ -1,0 +1,9 @@
+class Artwork
+  @@heading = 'Artworks'
+
+  def initialize(name, year, link)
+    @name = name
+    @year = year
+    @link = link
+  end
+end
