--- lib/html_reader/tempCodeRunnerFile.rb
@@ -1,1 +1,0 @@
-text
